fix: harden script execution contracts
